requirements.txt
chess
numpy
tensorflow
pillow
